pyproject.toml
[tool.poetry.scripts]
django = 'factotum.config.django.manage:main'

[tool.poetry]
name = "Factotum"
version = "0.1.0"
description = "Resource management software"
authors = ["Radosław Ganczarek <radoslaw@ganczarek.in>"]
license = "BSD"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
django-internationalflavor = "^0.4.1"

[tool.poetry.dev-dependencies]
ipython = "^7.23.0"
pylint = "^2.8.2"
mypy = "^0.812"
black = "^21.4b2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
